App refactor: Introduce FocusContextProvider
Improve App module's read- and compose-ability by moving FocusContext
concerns to a separate provider module.
